Fix year not automatically updating
`site.time` uses the time when the `jekyll` command was run, which I'm guessing is run every time I update the site. Since I hadn't updated this repo since 2019, the year was stuck in 2019. The template also weirdly seemed to use `20` followed by `%y` which appends the year without the century, when `%Y` just does the full year with the century.